renames reducers.ts to reducer.ts
