+++ components/App/index.ts
@@ -1,4 +1,4 @@
 export { default } from './App';
 export { useAppState, useAppDispatch, AppContextProvider } from './context';
-export * from './reducers';
+export * from './reducer';
 export * from './actions';
