+++ components/App/reducer.ts
