+++ components/System/index.ts
@@ -1,3 +1,3 @@
 export { systemActions } from './actions.ts';
 export { useSystemState, useSystemDispatch, SystemContextProvider } from './context';
-export { SystemReducer } from './reducers';
+export { SystemReducer } from './reducer.ts';
