+++ components/User/index.ts
@@ -1,3 +1,3 @@
 export { useUserDispatch, UserProvider } from './context';
 export { userActions } from './actions';
-export * from './reducers';
+export * from './reducer';
